Revert "EVG-16989 Upgrade legacy agent routes to v2 (#5811)"
This reverts commit d897981f9501f4e41b89c346db07ce61bdbbba67.
