+++ rest/data/task_test.go
@@ -473,7 +473,7 @@ func TestCheckTaskSecret(t *testing.T) {
 	}
 	code, err := CheckTaskSecret("task1", r)
 	assert.Error(err)
-	assert.Equal(http.StatusConflict, code)
+	assert.Equal(http.StatusUnauthorized, code)
 
 	r.Header.Set(evergreen.TaskSecretHeader, "abcdef")
 	code, err = CheckTaskSecret("task1", r)
