+++ service/api_middleware_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/evergreen-ci/evergreen/model/task"
 	"github.com/evergreen-ci/gimlet"
 	. "github.com/smartystreets/goconvey/convey"
-	"github.com/stretchr/testify/require"
 )
 
 func TestCheckHostWrapper(t *testing.T) {
@@ -32,10 +31,14 @@ func TestCheckHostWrapper(t *testing.T) {
 	queue := env.LocalQueue()
 
 	Convey("With a simple checkTask and checkHost-wrapped route", t, func() {
-		require.NoError(t, db.ClearCollections(host.Collection, task.Collection))
+		if err := db.ClearCollections(host.Collection, task.Collection); err != nil {
+			t.Fatalf("clearing db: %v", err)
+		}
 
 		as, err := NewAPIServer(env, queue)
-		require.NoError(t, err)
+		if err != nil {
+			t.Fatalf("creating test API server: %v", err)
+		}
 		var (
 			retrievedTask *task.Task
 			retrievedHost *host.Host
@@ -52,7 +55,9 @@ func TestCheckHostWrapper(t *testing.T) {
 		))).Get()
 
 		root, err := app.Handler()
-		require.NoError(t, err)
+		if err != nil {
+			t.Fatalf("creating test handler server: %v", err)
+		}
 
 		Convey("and documents representing a proper host-task relationship", func() {
 			So(t1.Insert(), ShouldBeNil)
@@ -60,7 +65,9 @@ func TestCheckHostWrapper(t *testing.T) {
 
 			w := httptest.NewRecorder()
 			r, err := http.NewRequest("GET", "/t1/", nil)
-			require.NoError(t, err)
+			if err != nil {
+				t.Fatalf("building request: %v", err)
+			}
 
 			Convey("a request without proper task fields should fail", func() {
 				root.ServeHTTP(w, r)
@@ -143,7 +150,9 @@ func TestCheckHostWrapper(t *testing.T) {
 
 			w := httptest.NewRecorder()
 			r, err := http.NewRequest("GET", "/t2/", nil)
-			require.NoError(t, err)
+			if err != nil {
+				t.Fatalf("building request: %v", err)
+			}
 
 			Convey("a request with proper task fields and host fields should fail", func() {
 				r.Header.Add(evergreen.TaskSecretHeader, t2.Secret)
@@ -162,10 +171,14 @@ func TestCheckHostWrapper(t *testing.T) {
 		})
 	})
 	Convey("With a requireTask and requireHost-wrapped route using URL params", t, func() {
-		require.NoError(t, db.ClearCollections(host.Collection, task.Collection))
+		if err := db.ClearCollections(host.Collection, task.Collection); err != nil {
+			t.Fatalf("clearing db: %v", err)
+		}
 
 		as, err := NewAPIServer(env, queue)
-		require.NoError(t, err)
+		if err != nil {
+			t.Fatalf("creating test API server: %v", err)
+		}
 
 		var (
 			retrievedTask *task.Task
@@ -184,7 +197,9 @@ func TestCheckHostWrapper(t *testing.T) {
 		))).Get()
 
 		root, err := app.Handler()
-		require.NoError(t, err)
+		if err != nil {
+			t.Fatalf("creating test API server: %v", err)
+		}
 
 		Convey("and documents representing a proper host-task relationship", func() {
 
@@ -193,7 +208,9 @@ func TestCheckHostWrapper(t *testing.T) {
 
 			w := httptest.NewRecorder()
 			r, err := http.NewRequest("GET", "/t1/h1", nil)
-			require.NoError(t, err)
+			if err != nil {
+				t.Fatalf("building request: %v", err)
+			}
 
 			Convey("a request with proper host params and fields should pass", func() {
 				r.Header.Add(evergreen.TaskSecretHeader, t1.Secret)
