+++ service/service.go
@@ -55,12 +55,10 @@ func GetRouter(as *APIServer, uis *UIServer) (http.Handler, error) {
 	rest := GetRESTv1App(as)
 
 	opts := route.HandlerOpts{
-		APIQueue:            as.queue,
-		QueueGroup:          as.env.RemoteQueueGroup(),
-		URL:                 as.Settings.Ui.Url,
-		GithubSecret:        []byte(as.Settings.Api.GithubWebhookSecret),
-		TaskDispatcher:      as.taskDispatcher,
-		TaskAliasDispatcher: as.taskAliasDispatcher,
+		APIQueue:     as.queue,
+		QueueGroup:   as.env.RemoteQueueGroup(),
+		URL:          as.Settings.Ui.Url,
+		GithubSecret: []byte(as.Settings.Api.GithubWebhookSecret),
 	}
 	route.AttachHandler(rest, opts)
 
@@ -72,12 +70,10 @@ func GetRouter(as *APIServer, uis *UIServer) (http.Handler, error) {
 	apiRestV2 := gimlet.NewApp()
 	apiRestV2.SetPrefix(evergreen.APIRoutePrefix + "/" + evergreen.RestRoutePrefix)
 	opts = route.HandlerOpts{
-		APIQueue:            as.queue,
-		QueueGroup:          as.env.RemoteQueueGroup(),
-		URL:                 as.Settings.Ui.Url,
-		GithubSecret:        []byte(as.Settings.Api.GithubWebhookSecret),
-		TaskDispatcher:      as.taskDispatcher,
-		TaskAliasDispatcher: as.taskAliasDispatcher,
+		APIQueue:     as.queue,
+		QueueGroup:   as.env.RemoteQueueGroup(),
+		URL:          as.Settings.Ui.Url,
+		GithubSecret: []byte(as.Settings.Api.GithubWebhookSecret),
 	}
 	route.AttachHandler(apiRestV2, opts)
